Better tests of dumping after set_types
When there is a 1-n mapping between oid and python classes, we might
pick a dumper that cannot really deal with the object to dump (for
instance trying to dump an int to numeric with the Decimal dumper).

Add a few dumpers which can deal with the different types and make sure
to register them so that the lookup by oid would use them. Because they
are slightly less efficient (with isinstance tests) we prefer to have
them only used by oid, instead of extending the domain of the dumpers
used in normal adaptation by type.

This commit only fixes the Python side. The C numeric dumpers require
more work.